完成教师添加页面需要的相关逻辑

+++ src/main/java/top/ostp/web/mapper/CollegeMapper.kt
@@ -20,7 +20,7 @@ interface CollegeMapper {
     fun selectByName(name: String): College?
 
     @Select("select * from college where id = #{id} limit 1")
-    fun selectById(id: Int): College?
+    fun selectById(id: Long): College?
 
     @Select("select * from college")
     @ResultType(College::class)
